Error testing
No error found in script, but added a cube to test error on trying to attach script to object to test if we can copy it. Getting error saying cannot attach editor script though. Doing research says I need to attach it to a monobehavior but not sure how that works or how to do that